+++ apps/elf-eval/tests/real_world_job_benchmark/manifest_summary.rs
@@ -1,0 +1,245 @@
+use serde_json::Value;
+
+pub(super) fn assert_external_adapter_manifest_summary(report: &Value) {
+	assert_eq!(
+		report.pointer("/external_adapters/schema").and_then(Value::as_str),
+		Some("elf.real_world_external_adapter_report/v1")
+	);
+	assert_eq!(
+		report.pointer("/external_adapters/manifest_id").and_then(Value::as_str),
+		Some(
+			"real-world-memory-project-adapters-2026-06-11-first-generation-continuity-source-store"
+		)
+	);
+	assert_eq!(
+		report.pointer("/external_adapters/docker_isolation/default").and_then(Value::as_bool),
+		Some(true)
+	);
+	assert_eq!(
+		report
+			.pointer("/external_adapters/docker_isolation/host_global_installs_required")
+			.and_then(Value::as_bool),
+		Some(false)
+	);
+	assert_eq!(
+		report.pointer("/external_adapters/summary/adapter_count").and_then(Value::as_u64),
+		Some(26)
+	);
+	assert_eq!(
+		report.pointer("/external_adapters/summary/external_project_count").and_then(Value::as_u64),
+		Some(19)
+	);
+	assert_eq!(
+		report.pointer("/external_adapters/summary/fixture_backed_count").and_then(Value::as_u64),
+		Some(1)
+	);
+	assert_eq!(
+		report
+			.pointer("/external_adapters/summary/live_baseline_only_count")
+			.and_then(Value::as_u64),
+		Some(6)
+	);
+	assert_eq!(
+		report.pointer("/external_adapters/summary/live_real_world_count").and_then(Value::as_u64),
+		Some(5)
+	);
+	assert_eq!(
+		report.pointer("/external_adapters/summary/research_gate_count").and_then(Value::as_u64),
+		Some(14)
+	);
+
+	assert_external_adapter_manifest_status_summary(report);
+	assert_external_adapter_manifest_scenario_summary(report);
+}
+
+fn assert_external_adapter_manifest_status_summary(report: &Value) {
+	assert_eq!(
+		report
+			.pointer("/external_adapters/summary/overall_status_counts/pass")
+			.and_then(Value::as_u64),
+		Some(4)
+	);
+	assert_eq!(
+		report
+			.pointer("/external_adapters/summary/overall_status_counts/wrong_result")
+			.and_then(Value::as_u64),
+		Some(6)
+	);
+	assert_eq!(
+		report
+			.pointer("/external_adapters/summary/overall_status_counts/lifecycle_fail")
+			.and_then(Value::as_u64),
+		Some(1)
+	);
+	assert_eq!(
+		report
+			.pointer("/external_adapters/summary/overall_status_counts/incomplete")
+			.and_then(Value::as_u64),
+		Some(0)
+	);
+	assert_eq!(
+		report
+			.pointer("/external_adapters/summary/overall_status_counts/blocked")
+			.and_then(Value::as_u64),
+		Some(10)
+	);
+	assert_eq!(
+		report
+			.pointer("/external_adapters/summary/overall_status_counts/not_encoded")
+			.and_then(Value::as_u64),
+		Some(5)
+	);
+	assert_eq!(
+		report
+			.pointer("/external_adapters/summary/capability_status_counts/mocked")
+			.and_then(Value::as_u64),
+		Some(1)
+	);
+	assert_eq!(
+		report
+			.pointer("/external_adapters/summary/capability_status_counts/unsupported")
+			.and_then(Value::as_u64),
+		Some(6)
+	);
+	assert_eq!(
+		report
+			.pointer("/external_adapters/summary/suite_status_counts/blocked")
+			.and_then(Value::as_u64),
+		Some(29)
+	);
+	assert_eq!(
+		report
+			.pointer("/external_adapters/summary/suite_status_counts/pass")
+			.and_then(Value::as_u64),
+		Some(27)
+	);
+	assert_eq!(
+		report
+			.pointer("/external_adapters/summary/suite_status_counts/incomplete")
+			.and_then(Value::as_u64),
+		Some(0)
+	);
+	assert_eq!(
+		report
+			.pointer("/external_adapters/summary/suite_status_counts/not_encoded")
+			.and_then(Value::as_u64),
+		Some(37)
+	);
+}
+
+fn assert_external_adapter_manifest_scenario_summary(report: &Value) {
+	assert_eq!(
+		report
+			.pointer("/external_adapters/summary/scenario_status_counts/real")
+			.and_then(Value::as_u64),
+		Some(0)
+	);
+	assert_eq!(
+		report
+			.pointer("/external_adapters/summary/scenario_status_counts/mocked")
+			.and_then(Value::as_u64),
+		Some(0)
+	);
+	assert_eq!(
+		report
+			.pointer("/external_adapters/summary/scenario_status_counts/unsupported")
+			.and_then(Value::as_u64),
+		Some(3)
+	);
+	assert_eq!(
+		report
+			.pointer("/external_adapters/summary/scenario_status_counts/blocked")
+			.and_then(Value::as_u64),
+		Some(24)
+	);
+	assert_eq!(
+		report
+			.pointer("/external_adapters/summary/scenario_status_counts/incomplete")
+			.and_then(Value::as_u64),
+		Some(5)
+	);
+	assert_eq!(
+		report
+			.pointer("/external_adapters/summary/scenario_status_counts/wrong_result")
+			.and_then(Value::as_u64),
+		Some(6)
+	);
+	assert_eq!(
+		report
+			.pointer("/external_adapters/summary/scenario_status_counts/lifecycle_fail")
+			.and_then(Value::as_u64),
+		Some(1)
+	);
+	assert_eq!(
+		report
+			.pointer("/external_adapters/summary/scenario_status_counts/pass")
+			.and_then(Value::as_u64),
+		Some(23)
+	);
+	assert_eq!(
+		report
+			.pointer("/external_adapters/summary/scenario_status_counts/not_encoded")
+			.and_then(Value::as_u64),
+		Some(13)
+	);
+	assert_eq!(
+		report
+			.pointer("/external_adapters/summary/scenario_position_counts/wins")
+			.and_then(Value::as_u64),
+		Some(10)
+	);
+	assert_eq!(
+		report
+			.pointer("/external_adapters/summary/scenario_position_counts/ties")
+			.and_then(Value::as_u64),
+		Some(11)
+	);
+	assert_eq!(
+		report
+			.pointer("/external_adapters/summary/scenario_position_counts/loses")
+			.and_then(Value::as_u64),
+		Some(1)
+	);
+	assert_eq!(
+		report
+			.pointer("/external_adapters/summary/scenario_position_counts/untested")
+			.and_then(Value::as_u64),
+		Some(53)
+	);
+	assert_eq!(
+		report
+			.pointer("/external_adapters/summary/scenario_outcome_counts/win")
+			.and_then(Value::as_u64),
+		Some(10)
+	);
+	assert_eq!(
+		report
+			.pointer("/external_adapters/summary/scenario_outcome_counts/tie")
+			.and_then(Value::as_u64),
+		Some(11)
+	);
+	assert_eq!(
+		report
+			.pointer("/external_adapters/summary/scenario_outcome_counts/loss")
+			.and_then(Value::as_u64),
+		Some(1)
+	);
+	assert_eq!(
+		report
+			.pointer("/external_adapters/summary/scenario_outcome_counts/not_tested")
+			.and_then(Value::as_u64),
+		Some(19)
+	);
+	assert_eq!(
+		report
+			.pointer("/external_adapters/summary/scenario_outcome_counts/blocked")
+			.and_then(Value::as_u64),
+		Some(29)
+	);
+	assert_eq!(
+		report
+			.pointer("/external_adapters/summary/scenario_outcome_counts/non_goal")
+			.and_then(Value::as_u64),
+		Some(5)
+	);
+}
